requirements.txt
aws-cdk-lib
constructs
boto3==1.37.15
Flask==3.0.2
Werkzeug==3.0.1
python-dotenv==1.0.1
